Move proposal messages functions to proposal/templates and made Discourse API call in background.
